+++ docs/source/conf.py
@@ -23,12 +23,12 @@ DESCRIPTION = (
 # -- Project information -----------------------------------------------------
 
 project = 'pyexcel'
-copyright = '2014-2020 Onni Software Ltd.'
+copyright = '2014-2022 Onni Software Ltd.'
 author = 'C.W.'
 # The short X.Y version
-version = '0.6.6'
+version = '0.7.0'
 # The full version, including alpha/beta/rc tags
-release = '0.6.7'
+release = '0.7.0'
 
 # -- General configuration ---------------------------------------------------