+++ pyexcel/docstrings/core.py
@@ -4,18 +4,23 @@
 
     Reusible docstrings for pyexcel.core
 
-    :copyright: (c) 2015-2020 by Onni Software Ltd.
+    :copyright: (c) 2015-2022 by Onni Software Ltd.
     :license: New BSD License
 """
 from . import keywords
 
-__GET_SHEET__ = (
-    keywords.EXAMPLE_NOTE_PAGINATION
-    + keywords.SOURCE_PARAMS_TABLE
+
+__GET_ARRAY__ = (
+    keywords.SOURCE_PARAMS_TABLE
     + """
 **Parameters**
     """
     + keywords.SOURCE_PARAMS
+)
+
+__GET_SHEET__ = (
+    keywords.EXAMPLE_NOTE_PAGINATION
+    + __GET_ARRAY__
 )
 
 __GET_BOOK__ = (
@@ -92,9 +97,9 @@ django model     dest_models, dest_initializers,
 
 GET_SHEET = __GET_SHEET__
 
-GET_ARRAY = __GET_SHEET__
+GET_ARRAY = __GET_ARRAY__
 
-IGET_ARRAY = __GET_SHEET__ + I_NOTE
+IGET_ARRAY = __GET_ARRAY__ + I_NOTE
 
 GET_DICT = __GET_SHEET__