+++ pyexcel/internal/garbagecollector.py
@@ -4,7 +4,7 @@
 
     Simple garbage collector
 
-    :copyright: (c) 2015-2020 by Onni Software Ltd.
+    :copyright: (c) 2015-2022 by Onni Software Ltd.
     :license: New BSD License
 """
 from pyexcel import docstrings as docs