+++ pyexcel/parser.py
@@ -4,7 +4,7 @@
 
     Extract tabular data from external file, stream or content
 
-    :copyright: (c) 2015-2020 by Onni Software Ltd.
+    :copyright: (c) 2015-2022 by Onni Software Ltd.
     :license: New BSD License
 """
 from pyexcel.internal import garbagecollector as gc