+++ pyexcel/plugins/parsers/__init__.py
@@ -4,7 +4,7 @@
 
     A list of built-in parsers
 
-    :copyright: (c) 2015-2020 by Onni Software Ltd.
+    :copyright: (c) 2015-2022 by Onni Software Ltd.
     :license: New BSD License
 """
 from pyexcel.plugins import PyexcelPluginChain