+++ pyexcel/renderer.py
@@ -4,7 +4,7 @@
 
     Renders pyexcel.Book and pyexcel.Sheet to any format
 
-    :copyright: (c) 2015-2020 by Onni Software Ltd.
+    :copyright: (c) 2015-2022 by Onni Software Ltd.
     :license: New BSD License
 """
 from pyexcel import _compact as compact
@@ -144,11 +144,7 @@ class BinaryRenderer(Renderer):
     """
     Renderer pyexcel data into a binary object
     """
-
-    def __init__(self, file_type):
-        Renderer.__init__(self, file_type)
-        if compact.PY3_AND_ABOVE:
-            self.WRITE_FLAG = "wb"
+    WRITE_FLAG = "wb"
 
     def get_io(self):
         io = compact.BytesIO()